refactor(EmergencyContactsScreen): mover estilos para arquivo styles.js
